Rename CompactionTask
